+++ gollybase/hlifealgo.cpp
@@ -1533,7 +1533,7 @@ default:       return "Illegal character in readmacrocell." ;
 	       p = line + 8 ;
 	       while (*p && *p <= ' ')
 		 p++ ;
-	       int cnt = atol(p) ;
+	       long cnt = atol(p) ;
 	       if (cnt < 0 || cnt > MAX_FRAME_COUNT)
 		  return "Bad FRAMES line" ;
 	       destroytimeline() ;
